requirements.txt
b2sdk
python-dotenv
yt-dlp
gallery-dl
apiflask[yaml]
flask-cors
flask-socketio
svix-ksuid
SQLAlchemy
gunicorn
boto3
pytest
